# public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/privacy-settings/create": {
      "post": {
        "operationId": "PrivacyController_createPrivacySettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePrivacySettingsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Privacy"
        ]
      }
    },
    "/v1/privacy-settings/update": {
      "put": {
        "operationId": "PrivacyController_updatePrivacySettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePrivacySettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Privacy"
        ]
      }
    },
...
  },
  "info": {
    "title": "XENAuth API",
    "description": "The official XENAuth API documentation",
    "version": "1.0",
    "contact": {
      "name": "Erzen Krasniqi",
      "url": "https://erzen.tk",
      "email": "erzenkrasniqi@matrics.io"
    },
    "termsOfService": "https://erzen.tk/terms",
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    }
  },
...
  "servers": [
    {
      "url": "https://localhost:3000",
      "description": "LOCAL server"
    },
    {
      "url": "https://api.erzen.xyz",
      "description": "DEV server"
    },
    {
      "url": "https://apis.erzen.xyz",
      "description": "PRODUCTION server"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdatePrivacySettingsDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  },
  "security": [
    {
      "access-token": []
    }
  ]
}